requirements.txt
Django==3.1.8
django-environ==0.4.5
psycopg2-binary==2.8.6
djangorestframework==3.12.4
httpie==2.4.0